+++ docs/CHANGELOG.md
@@ -19,12 +19,14 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * `amethyst_renderer::Rgba` is now a `Component` that changes the color and transparency of the entity
 it is attached to. ([#1282])
 * `AutoFov` and `AutoFovSystem` to adjust horizontal FOV to screen aspect ratio. ([#1281])
+* `AudioBundle.with_dj_system(...)` initializes the `DjSystem` as part of the `AudioBundle`. ([#1346])
 
 ### Changed
 
 * Make `application_root_dir` return a `Result<Path>` instead of a `String` ([#1213])
 * Remove unnecessary texture coordinates offset in `Sprite::from_pixel_values` ([#1267])
 * Changed `ActiveCamera` to have the `Option` inside. ([#1280])
+* `AudioBundle::new()` accepts no arguments and no longer initializes a `DjSystem`. ([#1346])
 
 ### Removed
 
@@ -44,6 +46,7 @@ it is attached to. ([#1282])
 [#1281]: https://github.com/amethyst/amethyst/pull/1281
 [#1302]: https://github.com/amethyst/amethyst/pull/1302
 [#1328]: https://github.com/amethyst/amethyst/pull/1328
+[#1346]: https://github.com/amethyst/amethyst/pull/1346
 
 ## [0.10.0] - 2018-12